+++ CHANGES.md
@@ -1,5 +1,7 @@
 
 
+
+- Add `Brr.El.Style.{top,left,right,bottom,position}`.
 - Fix `Blob.of_jstr`. It was not working. Thanks to Kiran Gopinathan for
   the report (#31).
 - `Ev.target_{of,to}_jv` take and return a `Jv.t` value instead of an `'a`.